requirements.txt
PyYAML==6.0
altair
matplotlib-inline==0.1.6
matplotlib==3.5.3
networkx==2.8.6
numpy==1.23.2
pandas==1.4.3
pydantic==1.9.2
scipy==1.9.0
seaborn==0.11.2
streamlit
gym
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cpu
